fix: more extractions
